refactor(Event): EventContainer added and pages/Events cleaned
